Unify to record-batch.{cpp,h,hpp}
