+++ c_glib/arrow-glib/Makefile.am
@@ -63,7 +63,6 @@ libarrow_glib_la_headers =			\
 	field.h					\
 	gobject-type.h				\
 	record-batch.h				\
-	record-batch-iterator.h			\
 	schema.h				\
 	table.h					\
 	table-builder.h				\
@@ -117,7 +116,6 @@ libarrow_glib_la_sources =			\
 	error.cpp				\
 	field.cpp				\
 	record-batch.cpp			\
-	record-batch-iterator.cpp		\
 	schema.cpp				\
 	table.cpp				\
 	table-builder.cpp			\
@@ -166,7 +164,6 @@ libarrow_glib_la_cpp_headers =			\
 	error.hpp				\
 	field.hpp				\
 	record-batch.hpp			\
-	record-batch-iterator.hpp		\
 	schema.hpp				\
 	table.hpp				\
 	table-builder.hpp			\
