+++ c_glib/arrow-glib/meson.build
@@ -30,7 +30,6 @@ sources = files(
   'error.cpp',
   'field.cpp',
   'record-batch.cpp',
-  'record-batch-iterator.cpp',
   'schema.cpp',
   'table.cpp',
   'table-builder.cpp',
@@ -86,7 +85,6 @@ c_headers = files(
   'field.h',
   'gobject-type.h',
   'record-batch.h',
-  'record-batch-iterator.h',
   'schema.h',
   'table.h',
   'table-builder.h',
@@ -141,7 +139,6 @@ cpp_headers = files(
   'error.hpp',
   'field.hpp',
   'record-batch.hpp',
-  'record-batch-iterator.hpp',
   'schema.hpp',
   'table.hpp',
   'table-builder.hpp',
